refactor: modifying the paragraphs constant
